% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_batch_data.R
\name{annotate_batch_data}
\alias{annotate_batch_data}
\title{A helper function}
\usage{
annotate_batch_data(path = path, results, HHID, RID, day, month, year,
  name, PID, game, order, seed = 1)
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function to check hue
}
