% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_enter_all.R
\name{auto_enter_all}
\alias{auto_enter_all}
\title{A helper function}
\usage{
auto_enter_all(path = "C:/Users/Mind Is Moving/Desktop/RICH/",
  pattern = ".jpg", start = 1, stop = 3, seed = 1, n_frames = 4,
  n_rows = 5, n_cols = 9, lower_hue_threshold = 210,
  upper_hue_threshold = 230, colors = c("empty", "red"), img, locs,
  focal = "CTR", case, thresh = 0.25, ordered = NULL,
  lower_saturation_threshold = 0.05, lower_luminance_threshold = 0.05,
  upper_luminance_threshold = 0.95, border_size = 8, iso_blur = 2)
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function
}
