% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all_random_payouts.R
\name{build_all_random_payouts}
\alias{build_all_random_payouts}
\title{Select random payouts and generate corresponding}
\usage{
build_all_random_payouts(
  path,
  n_alloc_keep = 1,
  n_alloc_receive = 1,
  guess_margin = 0,
  guess_payout_amt = 0,
  allocations_df = NA
)
}
\arguments{
\item{n_alloc_keep}{Number of 'outgoing' allocations to select:
will implement the amount 'kept' by the player. If there are fewer
allocations than this number, implement the closest possible figure.}

\item{n_alloc_receive}{Number of 'incoming' allocations to select:
will implement the amount 'give' to the player. If there are fewer
allocations than this number, implement the closest possible figure.}

\item{guess_margin}{the margin within which an expected allocation
will be rewarded. E.g. set to 10 to reward all guesses within ±10
of actual allocation.}

\item{guess_payout_amt}{The amount given per correct guess.}
}
\description{
Randomly select a number of allocations made by and to each player,
and generate the json files required for the payout activity in
DieTryinSR.
}
\details{
This function requires the results of the allocations to have been
compiled with compile_subset_survey_data().
}
