% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all_revelations.R
\name{build_all_revelations}
\alias{build_all_revelations}
\title{Build all payout files for DieTryinSR in enhanced mode}
\usage{
build_all_revelations(
  path,
  remove_optouts = TRUE,
  gid_size = 4,
  allocations_df = NA
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{allocations_df}{A dataframe with all the allocations that need to be revealed.}

\item{keep_anonymous}{Set to TRUE if you want to generate files for the games that were played in the "anonymous" condition. Defaults to FALSE.}
}
\description{
This is a simple loop that takes game results from an initial set of dicator games (played in "enhanced" mode) and generates files that can be used for the payout functionality in DieTryinSR. This should only be used after the SubsetContributions-SummaryTable.csv file is created, using compile_subset_survey_data().
}
