% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_roster.R
\name{build_roster}
\alias{build_roster}
\title{A function to automatically build a photo roster for data collection}
\usage{
build_roster(
  path,
  pattern = ".jpg",
  start = 1,
  stop = 3,
  seed = 1,
  n_panels = 4,
  n_rows = 5,
  n_cols = 8,
  ordered = NULL,
  size_out = 1000,
  chip_width = 300,
  asr = 1.618,
  width = 2.5
)
}
\arguments{
\item{path}{Full path to main folder.}

\item{pattern}{File extension of photos. Should be ".jpg" or ".JPG".}

\item{start}{Location of start of PID in file name. If files are saved as "XXX.jpg" for example, this is 1.}

\item{stop}{Location of end of PID in file name. If files are saved as "XXX.jpg" for example, this is 3.}

\item{seed}{A seed for the random number generator to sort the order of photos in the array. This should match the seed used to make the survey.}

\item{n_panels}{Number of frames/panels/blocks of photos to be output. I use four big panels and randomize order at each game.}

\item{n_rows}{Number of rows per panel. With 7cm x 10cm photos, I use five rows of photos per panel.}

\item{n_cols}{Number of rows per panel. With 7cm x 10cm photos, I use six to eight cols of photos per panel.}

\item{ordered}{A list of IDs if photograph order is to be explicilty coded. This overwrites random sorting, otherwise NULL.}

\item{size_out}{Width of standardized photos (pixels).}

\item{chip_width}{Width at which reference chip should be printed (pixels).}

\item{asr}{Aspect ratio of each photo.}

\item{width}{Width at which each photo on the poster should be printed (inches).}
}
\description{
This is a helper function to plot the roster. Posters are saved in the "Survey" folder.
}
\examples{
\dontrun{
 build_roster(path, pattern=".jpg", start=1, stop=3, n_panels=4, n_rows=6, n_cols=14, 
             size_out=1000, chip_width=300, asr=1.6180, width = 2.5, seed=1, ordered = NULL )
                   }
}
