% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_subset_allocation.R
\name{build_subset_allocation}
\alias{build_subset_allocation}
\title{Build a single allocation evaluation file for DieTryinSR in enhanced mode}
\usage{
build_subset_allocation(
  path,
  subdir = "SubsetContributions",
  pid = NULL,
  aid = NULL,
  seed = 123,
  gid_size = 4,
  condition = "",
  endowment = 100,
  opt_out_keep = 0,
  ask_opt_out = FALSE
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pid}{ID code of participant.}

\item{aid}{ID of alter (about whom reputational eval is being made)}

\item{seed}{Number to use in seeding the randomization process if used.}

\item{gid_size}{Number of terms in hashcodes for the game IDs.}
}
\description{
Build a single allocation evaluation file for DieTryinSR in enhanced mode
}
