% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_subset_expectation.R
\name{build_subset_expectation}
\alias{build_subset_expectation}
\title{Build a single payoff file for DieTryinSR in enhanced mode}
\usage{
build_subset_expectation(
  path,
  subdir = "SubsetExpectations",
  pid = NULL,
  aid = NULL,
  offer = NA,
  anonymous = FALSE,
  seed = 123,
  gid_size = 4
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pid}{ID code of receiver (who will be presented the payout made by giver)}

\item{aid}{ID of giver (whose offer is already known)}

\item{offer}{Offer that was made by giver to receiver.}

\item{seed}{Number to use in seeding the randomization process if used.}

\item{gid_size}{Number of terms in hashcodes for the game IDs.}
}
\description{
Build a single payoff file for DieTryinSR in enhanced mode
}
