% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_subset_payout.R
\name{build_subset_payout}
\alias{build_subset_payout}
\title{Build a single payout file for DieTryinSR in enhanced mode}
\usage{
build_subset_payout(
  path,
  subdir = "SubsetPayouts",
  pid = NULL,
  amt_kept = 0,
  amt_received = 0,
  n_kept = 0,
  n_received = 0,
  guess_margin = 0,
  guess_payout_amt = 0
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pid}{ID code of participant.}

\item{aid}{ID of alter}

\item{pid_role}{Set to "receive" if this was an amount given to pid, "keep" if this was an amount kept by pid.}

\item{amt}{Amount given or kept}

\item{anonoymous}{Set to TRUE to hide alter's photograph (in practice should only be used with pid_role == receive)}

\item{opted_out}{Set to TRUE if giver opted out. Only useful for pid_role == "keep". amt still needs to be accurate}

\item{seed}{Number to use in seeding the randomization process if used.}

\item{gid_size}{Number of terms in hashcodes for the game IDs.}
}
\description{
Build a single payout file for DieTryinSR in enhanced mode
}
