% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_subset_rep_eval.R
\name{build_subset_rep_eval}
\alias{build_subset_rep_eval}
\title{Build a single reputational evaluation file for DieTryinSR in enhanced mode}
\usage{
build_subset_rep_eval(
  path,
  subdir = "SubsetRep1",
  pid = NULL,
  aid = NULL,
  seed = 123,
  gid_size = 4,
  questions = c("question1", "question2", "question3"),
  likert_levels = c("Nope", "Meh", "Kinda", "Yeah", "Mega"),
  dont_know_text = "dunno"
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pid}{ID code of participant.}

\item{aid}{ID of alter (about whom reputational eval is being made)}

\item{seed}{Number to use in seeding the randomization process if used.}

\item{gid_size}{Number of terms in hashcodes for the game IDs.}
}
\description{
Build a single reputational evaluation file for DieTryinSR in enhanced mode
}
