% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_subset_surveys.R
\name{build_subset_surveys}
\alias{build_subset_surveys}
\title{Build photo-roster subset surveys (e.g., for RICH-PGG games.)}
\usage{
build_subset_surveys(
  path,
  pid = NULL,
  id_set = NULL,
  game_name = "Choice",
  entry_type = "random",
  set_size = 4,
  pattern = ".jpg",
  token_color = "navyblue",
  max_iter = 10000,
  full_alter_set = TRUE,
  height = 8.5,
  width = 11,
  seed = NA,
  gid_size = 4
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pid}{ID code of focal recipient.}

\item{id_set}{IDs of alter recipients (as vector).}

\item{game_name}{Used to label PDF surveys.}

\item{entry_type}{Pull from all files in SubsetData folder using: "random", "choice", or "predefined". 
If entry_type is "random", all alters will be selected randomly, in such a way as to fairly balance the number of games 
in which each alter appears.
If entry_type is "choice", then alters will be defined as in the partner-choice game data.
If entry_type is "predefined", then a specific "pid" and "id_set" can be added to the inputs in order to build a 
single PDF file with a specific predefined set of alters file. This can be used, for example, if the researcher wants
to manipulate the roster to feature alters with specific characteristics.}

\item{set_size}{Size of set of possible alters.}

\item{pattern}{Should "JPG" be "jpg" file endings be used to load photos?}

\item{token_color}{Number or "color" of tokens placed. If data was entered manually, place whatever value was stored 
to indiciate a tie: "1"}

\item{max_iter}{Max time to search for full-scope legal permutations. In random mode, we seek to ensure that each
alter appears in exactly the same number of games. Thus, we iterate randomizations until we find such a set.}

\item{full_alter_set}{Include all alters on roster, or only those who played first round of game?}

\item{height}{Size of PDF output.}

\item{width}{Size of PDF output.}

\item{seed}{Number to use in seeding the randomization process if used.}

\item{gid_size}{Number of characters in hashcodes for the game IDs.}
}
\value{
A file folder, SubsetSurveys, filled with PDFs of sub-surveys to run, and a second folder, SubsetContributions, 
 filled with CSV files that will be used to record survey results.
}
\description{
This is a wraper function to create suveys (PDFs) to collect data for RICH-PGG contributions and similar games.
}
\examples{
\dontrun{
build_subset_surveys(path, pattern = ".jpg", token_color="navyblue", full_alter_set = TRUE,
                      entry_type="random", max_iter=10000, set_size=4, 
                      height=8.5, width=11, seed=123, gid_size=4)
}

}
