% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_unified_database.R
\name{build_unified_database}
\alias{build_unified_database}
\title{A helper function to merge DieTryin databases collected on different devices.}
\usage{
build_unified_database(
  main_directory,
  sub_directory = "Day1Backup",
  android_device_names,
  temp_directory = "temporary_RICH_files"
)
}
\arguments{
\item{main_directory}{Path to the directory containing the RICH folder.}

\item{sub_directory}{Which sub-directory will supply the .tar databases to be merged?}

\item{android_device_names}{What are the device names of the Androids used to collect these data? Use find_android_devices(main_directory)$devices_clean to find out, if devices are connected.}

\item{temp_directory}{What name should be used to hold temporary files? Doesn't really matter.}
}
\value{
This function dose not return any objects, but it creates a clone of the RICH database called RICH_Merged inside of main_directory.
}
\description{

}
