% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_classification.R
\name{check_classification}
\alias{check_classification}
\title{A function to check automatic classifications}
\usage{
check_classification(path, dat, n_frames = 2, n_rows = 4, n_cols = 5,
  focal = "BS1", case = "Friend")
}
\arguments{
\item{path}{Full path to main folder.}

\item{dat}{Classification output.}

\item{n_frames}{Number of frames/panels/blocks of photos to be output. I use four big panels and randomize their order at each game.}

\item{n_rows}{Number of rows per panel. With 7cm x 10cm photos, I use five rows of photos per panel.}

\item{n_cols}{Number of cols per panel. With 7cm x 10cm photos, I use six to eight cols of photos per panel.}

\item{focal}{ID code of focal individual.}

\item{case}{ID code for game/case/question: e.g., "GivingGame".}
}
\description{
This is a helper function to plot infered connections. Run this for each game, and it will plot predicted tokens on the real images. If model performance
is not good, then tweak parameters until classification is accurate. Photos are saved in the "ClassifiedPhotos" folder.
}
\examples{
\dontrun{
 check_classification(path=path, dat[[1]], n_frames = 2, n_rows=4, n_cols=5, focal="SK1", case="FriendshipsData")
                   }
}
