% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_colors.R
\name{clean_colors}
\alias{clean_colors}
\title{A function to remove token colors from standardized photos}
\usage{
clean_colors(
  path = path,
  id = "blank",
  id_full = NA,
  id_short = NA,
  lower_hue_threshold = c(120, 270, 185),
  upper_hue_threshold = c(180, 330, 210),
  rotation_angle = c(-65, 70, 40),
  mode = "test"
)
}
\arguments{
\item{path}{Full path to main folder.}

\item{id}{Unique ID of focal individual/respondent. Used if user wants to test colors by setting mode="test".}

\item{id_full}{Full path to photograph. Used by remove_color_collisions function. Set to NA otherwise.}

\item{id_short}{ID code plus .jpg extension. Used by remove_color_collisions. Set to NA otherwise.}

\item{lower_hue_threshold}{A vector of lower hue thresholds for each token color. To use three token colors, instead of the single token in the defaults, use: e.g., c(120, 210, 330).}

\item{upper_hue_threshold}{A vector of upper hue thresholds for each token color. To use three token colors, instead of the single token in the defaults, use: e.g., c(150, 250, 355).}

\item{rotation_angle}{A vector of angles with which to shift colors inside of the token range. Values are specified in degree: e.g., c(-65, 70, 40). Calculations are made mod 360.}

\item{mode}{Set to "test", to check photos by hand. The "save" option is used only by remove_color_collisions.}
}
\description{
This function allows you to make automatic data entry more accurate by filtering out token color from the photo roster. Simply set a path to the main folder. 
Then supply an ID code and the token color vectors and shift angles. It is recomended to use cool token colors (e.g., green, blue, and purple), and shift all
token colors in the standized photos towards warmer yellow, red, and orange hues.
}
