% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_android_settings.R
\name{compile_android_settings}
\alias{compile_android_settings}
\title{Build a map between GIDs and PIDs}
\usage{
compile_android_settings(
  path,
  photoMode = "onlyfocal",
  entryMode = "permanent",
  photoNumber = "twolines"
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{photoMode}{For games like the PGG or SnowDrift, where the focal persons can either give or not, set to "onlyfocal". For games were the focal
can give to anyone on the roster, set to "fullset".}

\item{entryMode}{Set to "permanent" so that data cant be changed once saved. Set to "temporary" if saved data are allowed to be overwritten.}

\item{photoNumber}{The app allows up to 9 total targets when set to "threelines", and 6 targets when set to "twolines". Three people per line.}
}
\description{
This is a small helper function to create suveys JSON files that guide the app behavior.
}
