% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_data}
\alias{compile_data}
\title{A function to compile CSV data from RICH economic games into an edgelist and summary table}
\usage{
compile_data(path = path, game = "GivingData", batch = FALSE)
}
\arguments{
\item{path}{Full path to main folder.}

\item{game}{Select the data to merge. "GivingData" for Giving. "LeavingData" for Leaving. "ReducingData" for Reducing.}
}
\description{
This function allows you to merge the individual-specific CSVs. Simply set a path to the main folder. Then run the function. See details below.
}
\examples{
\dontrun{
  compile_data(path=path,game="GivingData")
                   }
}
