% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_subset_survey_data.R
\name{compile_subset_survey_data}
\alias{compile_subset_survey_data}
\title{A function to compile individual CSVs in the SubsetContributions folder into a flat database.
Used to create a database after each individual record is entered.}
\usage{
compile_subset_survey_data(path = path, format = "json")
}
\arguments{
\item{path}{Path to RICH folder.}

\item{format}{Is the data saved as csv or json?}
}
\value{
A CSV of compiled contribution records will be added to the Results folder.
}
\description{
A function to compile individual CSVs in the SubsetContributions folder into a flat database.
Used to create a database after each individual record is entered.
}
\examples{
\dontrun{
compile_subset_survey_data(path)
}

}
