% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copydir.R
\name{copydir}
\alias{copydir}
\title{A helper function}
\usage{
copydir(from, to)
}
\arguments{
\item{from}{An example directory, to be cloned.}

\item{to}{Where to saved the cloned layout.}
}
\value{
This function dose not return any objects.
}
\description{
This is a helper function to clone directory structures.
}
