% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_to_android_devices.R
\name{deploy_to_android_devices}
\alias{deploy_to_android_devices}
\title{Deploy a DieTryin datebase to all connected Android devices.}
\usage{
deploy_to_android_devices(
  main_directory,
  build_tar = TRUE,
  auto_untar = TRUE,
  remove_tar = TRUE
)
}
\arguments{
\item{main_directory}{Path to the directory containing the RICH folder.}

\item{build_tar}{Boolean. Does the database need to be compressed with tar? Set to TRUE unless a valid tar object already exists.}

\item{auto_untar}{Boolean. Should DieTryin initiate tar decompression upon uploading to each device? Its simplest to set to TRUE. However, it may be faster to set 'auto_untar=FALSE' here and untar directly in the Android app using the RAR GUI.}

\item{remove_tar}{Boolean. Should the temporary .tar file be removed after the database is decompressed. It is best to keep the Androids clean, and set to TRUE.}
}
\value{
Nothing is returned by this function. However, the full RICH database is pushed up to all connected Android devices.
}
\description{
This is a helper function that will push a created JSON database to all connected Android phones. 
Later, the data from each Android can be merged to a single unified database.
}
