% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DieTryin.R
\docType{package}
\name{DieTryin}
\alias{DieTryin}
\title{DieTryin: Social network collection and automated entry using R}
\description{
DieTryin is an R package designed to facilitate the collection of roster-based network data, 
and to run network-structured economic games—such as Recipient Identity-Conditioned Heuristics (RICH) games.
}
