% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsize.R
\name{downsize}
\alias{downsize}
\title{Batch resize}
\usage{
downsize(path = path, scaler = 2)
}
\arguments{
\item{path}{path Full path to main folder.}

\item{scaler}{Factor by which image will be shrunk.}
}
\description{
This is a helper function to batch resize raw images of game boards. Set scaler=1 to simply pass images from the raw folder into the output file folder with no scaling.
}
