% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_data.R
\name{enter_data}
\alias{enter_data}
\title{A function to enter data from RICH economic games}
\usage{
enter_data(path = path, pattern = ".jpg", start = 1, stop = 3,
  seed = 1, frames = 4, rows = 5, cols = 8)
}
\arguments{
\item{path}{Full path to main folder.}

\item{pattern}{File extension of photos. Should be .jpg or .JPG.}

\item{start}{Location of start of PID in file name. If files are saved as XXX.jpg for example, this is 1.}

\item{stop}{Location of end of PID in file name. ... and this is 3.}

\item{seed}{A seed for the random number generator to sort the order of photos in the array. This should match the seed used to make the survey.}

\item{frames}{Number of frames/panels/blocks of photos to be output. I use four big panels and randomize order at each game.}

\item{rows}{Number of rows per panel. With 7cm x 10cm photos, I use five rows of photos per panel.}

\item{cols}{Number of rows per panel. With 7cm x 10cm photos, I use six to eight cols of photos per panel.}
}
\description{
This function allows you to speed up data entry. Simply set a path to the main folder. Then run the function. In the popup, enter "Y" if the data is for a new person, and "N" if it is for a second game from the same person. In the header for each data.entry table, you must enter 'G', 'L', or 'R' to indicate which game: Giving, Leaving, or Reducing, the data correspond to. In the function call, set the number of panels, and the number of rows and cols per panel, as in the build_survey file. Then run the function. This function relies on the 'data.entry' function. This is only supported on some builds. I've only tested on Windows. See details below.
}
\examples{
\dontrun{
  enter_data(path=path,pattern=".jpg",start=1,stop=3,seed=1,frames=4,rows=5,cols=8)
                   }
}
