% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_subset_surveys.R
\name{enter_subset_surveys}
\alias{enter_subset_surveys}
\title{A function to enter subset survey data from games like the RICH PGG.
Useful for data entry when a restricted sub-set of the photo roster is used in economic games.}
\usage{
enter_subset_surveys(path)
}
\arguments{
\item{path}{Path to RICH folder.}
}
\value{
The template CSV files in the SubsetContributions folder will be updated to include 
data on contribution levels.
}
\description{
A function to enter subset survey data from games like the RICH PGG.
Useful for data entry when a restricted sub-set of the photo roster is used in economic games.
}
\examples{
\dontrun{
enter_subset_surveys(path)
}

}
