% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{extractor}
\alias{extractor}
\title{A helper function}
\usage{
extractor(image, locations, n_rows = 3, n_cols = 5, cell_scale = 70,
  cell_width = 1, cell_height = 1.5)
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function to unskew images
}
