% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_android_devices.R
\name{find_android_devices}
\alias{find_android_devices}
\title{Find all connected Android devices}
\usage{
find_android_devices(main_directory)
}
\arguments{
\item{main_directory}{Path to the directory containing the RICH folder.}
}
\value{
A list containing basic details about the connected Android devices.
}
\description{
This is a helper function used to find Android devices connected to the computer, check that permissions are granted, and pull device names.
}
