% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hue.R
\name{get_hue}
\alias{get_hue}
\title{get_hue}
\usage{
get_hue(path)
}
\arguments{
\item{path}{Full path to image. file.choose() is often useful. Example: get_hue(file.choose())}
}
\description{
This is a helper function to identify HSL hue values. ust set a path to an image, and click to get a pixel color.
}
