% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_points_automatic.R
\name{grab_points_automatic}
\alias{grab_points_automatic}
\title{grab_points_automatic}
\usage{
grab_points_automatic(
  path_imgs,
  path_refset,
  d_x = 10,
  d_y = 10,
  plot_corners = TRUE
)
}
\arguments{
\item{path_imgs}{the path to the images to be scaned for corners.}

\item{path_refset}{the path to the reference chips.}

\item{d_x}{x width of kernel used to scan for corners.}

\item{d_y}{y width of kernel used to scan for corners.}

\item{plot_corners}{should predicted corners be plotted in situ?}
}
\value{
A list of lists of length 4. Each slot contains the locations of game-board corners.
}
\description{
This is a helper function.
}
