% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_to_LeavingData.R
\name{L_to_LeavingData}
\alias{L_to_LeavingData}
\title{L_to_LeavingData}
\usage{
L_to_LeavingData(path, max_coins, token_color)
}
\arguments{
\item{path}{The base file directory.}

\item{max_coins}{The maximum number of coins to be placed.}

\item{token_color}{The color label of the tokens used in the leaving game.}
}
\value{
Nothing is returned by this function, but data from the folder "L" are duplicated to "LeavingData".
}
\description{
This is a helper function to convert automatically entered token placements in the leaving game into
numeric data, as if it were manually entered.
}
