% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_align_string.R
\name{make_align_string}
\alias{make_align_string}
\title{A helper function}
\usage{
make_align_string(x)
}
\arguments{
\item{x}{An object.}
}
\description{
This is just a helper function.
}
\examples{
\dontrun{
  make_align_string(x)
                   }
}
