% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_data_to_subset_designer.R
\name{pass_data_to_subset_designer}
\alias{pass_data_to_subset_designer}
\title{A function to pass partner choice data to the subset suvery builder tool.
Useful for the RICH partner choice and public goods game.}
\usage{
pass_data_to_subset_designer(path, folder)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{folder}{Which folder holds the partner choice data? Maybe "PartnerChoice".}
}
\value{
The data files in "folder", will be cloned into "SubsetData", where they can be read 
 by the survey tool builder functions.
}
\description{
A function to pass partner choice data to the subset suvery builder tool.
Useful for the RICH partner choice and public goods game.
}
\examples{
\dontrun{
pass_data_to_subset_designer(path, folder="PartnerChoice")
}

}
