% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_from_android_devices.R
\name{pull_data_from_android_devices}
\alias{pull_data_from_android_devices}
\title{A helper function to pull .tar data objects from all connected Androids}
\usage{
pull_data_from_android_devices(
  main_directory,
  build_tar = TRUE,
  remove_tar = TRUE,
  sub_directory = NULL
)
}
\arguments{
\item{main_directory}{Path to the directory containing the RICH folder.}

\item{build_tar}{Should a new .tar file be created on the Android? This is almost always needed, unless the user has used a third-party app to create the .tar compressed version of the RICH database, with name RICHdata.tar.}

\item{remove_tar}{Should the RICHdat.tar file, used to send data back to R, be deleted after the data is pulled? Recommened to keep the Android clean.}

\item{sub_directory}{Please supply a name for the new subdirectory where the .tar files will be saved. For example: sub_directory="Day1Backup". We recommend backing up all devices each day, and saving all copies.}
}
\value{
This function does not return any objects.
}
\description{
This is a helper function to pull data from all connected Androids, saving each version as an independent .tar file, indexed by the Android's device ID.
Multiple database files can be automatically merged using other functions in DieTryin.
}
