% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_file_to_android_devices.R
\name{push_file_to_android_devices}
\alias{push_file_to_android_devices}
\title{Deploy a file to all connected Android devices.}
\usage{
push_file_to_android_devices(main_directory, file_path)
}
\arguments{
\item{main_directory}{Path to the directory containing the RICH folder.}

\item{file_path}{Path to file.}
}
\value{
Nothing is returned by this function. However, the full RICH database is pushed up to all connected Android devices.
}
\description{
A small function to send a file to all connected Androids. 
Useful for pushing the DieTryinSR and DieTryinCam APK files to new devices.
}
