% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_something.R
\name{say_something}
\alias{say_something}
\title{say_something}
\usage{
say_something(message, voice)
}
\arguments{
\item{message}{The message to read.}

\item{voice}{The voice to use.}
}
\description{
This is a helper function.
}
