% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setup_folders}
\alias{setup_folders}
\title{A function to create directory}
\usage{
setup_folders(path = path, add = NULL)
}
\arguments{
\item{path}{Full path to where folder will be stored.}
}
\description{
This function allows you to build the directories for the data
}
\examples{
\dontrun{
  path<-"C:\\\\Users\\\\BirdySanders\\\\Desktop"
  setup_folders(path)
                   }
}
