% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shatter.R
\name{shatter}
\alias{shatter}
\title{A helper function}
\usage{
shatter(image, locations, n_rows = 5, n_cols = 9,
  lower_hue_threshold = lower_hue_threshold,
  upper_hue_threshold = upper_hue_threshold,
  lower_saturation_threshold = lower_saturation_threshold,
  lower_luminance_threshold = lower_luminance_threshold,
  upper_luminance_threshold = upper_luminance_threshold,
  border_size = border_size, iso_blur = iso_blur)
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function to break array into cells and check hue differences
}
