% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_survey_compiler_partner_choice.R
\name{subset_survey_compiler_partner_choice}
\alias{subset_survey_compiler_partner_choice}
\title{Build RICH subset surveys (e.g., for partner choice or PG games.)}
\usage{
subset_survey_compiler_partner_choice(
  path,
  pattern = ".jpg",
  token_color = "navyblue",
  set_size = 4,
  height = 8.5,
  width = 11,
  seed = 123,
  gid_size = 4,
  game_name = "Choice",
  RID = "CR",
  day = 1,
  month = 3,
  year = 2023
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pattern}{Should "JPG" be "jpg" be used to load photos?}

\item{token_color}{"color" of tokens placed. If data was entered manually, place whatever value was stored 
to indiciate a tie: e.g., "1"}

\item{set_size}{Size of set of possible alters.}

\item{height}{Size of PDF output.}

\item{width}{Size of PDF output.}

\item{seed}{Number to use in seeding randomizer. Not applicable with choice data.}

\item{gid_size}{Number of characters in hashcodes for the game IDs.}

\item{game_name}{Used to label PDF surveys.}

\item{RID}{ID of researcher.}

\item{day}{Day of interview.}

\item{month}{Month of interview.}

\item{year}{Year of interview.}
}
\value{
A file folder, SubsetSurveys, filled with PDFs of sub-surveys to run, and a second folder, SubsetContributions, 
 filled with CSV files that will be used to record survey results.
}
\description{
This is a small helper function to create suveys (PDFs) to collect data for PGG contributions and similar games.
}
\examples{
\dontrun{
subset_survey_compiler_partner_choice(path, pattern = ".jpg", token_color="navyblue", set_size=4, 
                                 height=8.5, width=11, seed=123, gid_size=4)
}

}
