% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_survey_compiler_random.R
\name{subset_survey_compiler_random}
\alias{subset_survey_compiler_random}
\title{Build RICH subset surveys (e.g., for partner choice or PG games.)}
\usage{
subset_survey_compiler_random(
  path,
  pattern = ".jpg",
  token_color = "navyblue",
  set_size = 4,
  height = 8.5,
  width = 11,
  seed = 123,
  max_iter = 10000,
  gid_size = 4,
  game_name = "Random"
)
}
\arguments{
\item{path}{Path to RICH folder.}

\item{pattern}{Should "JPG" be "jpg" be used to load photos?}

\item{token_color}{"color" of tokens placed. If data was entered manually, place whatever value was stored 
to indiciate a tie: e.g., "1"}

\item{set_size}{Size of set of possible alters.}

\item{height}{Size of PDF output.}

\item{width}{Size of PDF output.}

\item{seed}{Number to use in seeding the randomizer.}

\item{max_iter}{Max time to search for legal permutations.}

\item{gid_size}{Number of characters in hashcodes for the game IDs.}

\item{game_name}{Used to label PDF surveys.}
}
\value{
A file folder, SubsetSurveys, full of PDFs of subset-surveys to run, and a second folder,
 SubsetContributions, full of CSV files to store results.
}
\description{
This is a small helper function to create suveys (PDFs) to collect data for PGG contributions and similar games.
}
\examples{
\dontrun{
subset_survey_compiler_random(path, pattern = ".jpg", token_color="navyblue", set_size=4, 
                                 height=8.5, width=11, seed=123, gid_size=4, max_iter=10000)
}

}
