% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{A helper function}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{A string.}

\item{n}{Number of characters to return, from right.}
}
\value{
A substring. For example, substrRight("patiently waiting.mp3", 3) returns "mp3".
}
\description{
This is a helper function to grab file extensions.
}
