% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wipe_RICH_files_from_android_devices.R
\name{wipe_RICH_files_from_android_devices}
\alias{wipe_RICH_files_from_android_devices}
\title{A helper function to clean the Androids of database files}
\usage{
wipe_RICH_files_from_android_devices(
  main_directory,
  wipe_RICH = FALSE,
  wipe_tar_payload = FALSE,
  wipe_tar_database = FALSE
)
}
\arguments{
\item{main_directory}{Path to the directory containing the RICH folder.}

\item{wipe_RICH}{Should the RICH database directory be deleted? Be very careful with this one! Data is permanently deleted, and this may be the only copy!}

\item{wipe_tar_payload}{Should the payload DieTryinPayload.tar file, used to send the database to each Android, be deleted?}

\item{wipe_tar_database}{Should the RICHdat.tar file, used to send data back to R, be deleted?}
}
\value{
This function does not return any objects.
}
\description{
This is a helper function to clean up the Androids. Three kinds of files can be removed. The payload "DieTryinPayload.tar" file used to send the database to each Android. The database directory itself. And, the "RICHdata.tar" file, used to send data back to R.
Be careful when using, as data loss is permanent. The user must set the arguments purposefully. A second set of prompts is used to comfirm, before deletion occurs.
}
