% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable_custom.R
\name{xtable_custom}
\alias{xtable_custom}
\title{A helper function}
\usage{
xtable_custom(x, ...)
}
\arguments{
\item{x}{An object.}
}
\description{
This is just a helper function.
}
\examples{
\dontrun{
  xtable_custom(x)
                   }
}
