% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{annotate_data}
\alias{annotate_data}
\title{A function to append header data to token allocation data entered via a scan of photographs}
\usage{
annotate_data(
  path,
  results,
  HHID,
  RID,
  day,
  month,
  year,
  name,
  ID,
  game,
  order,
  seed
)
}
\arguments{
\item{path}{Full path to main folder.}

\item{results}{A cell from the results list exported by auto_enter_all.}

\item{HHID}{Household ID of focal individual/respondent.}

\item{RID}{ID of researcher.}

\item{day}{Day of interview.}

\item{month}{Month of interview.}

\item{year}{Year of interview.}

\item{name}{Name of focal individual/respondent.}

\item{ID}{Unique ID of focal individual/respondent.}

\item{game}{ID for case/game/question.}

\item{order}{Order of frames/panels of photos as presented to the respodent: e.g., with 4 frames, "ABCD", "CDBA", etc. are legal entries.}

\item{seed}{A seed for the random number generator to sort the order of photos in the array. This should match the seed used to make the survey.}
}
\description{
This function appends header data to automatically entered data. The end result is a "csv" with the same structure as that
which is exported by the standard enter_data function.
}
\examples{
\dontrun{
annotate_data(path = path, results=Game_all5, HHID="BPL", RID="CR", day=12, month=4, year=2020, 
               name = "Cody", PID="CTR", game="TrustData", order="BA", seed = 1)
                   }
}
