% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_enter_data.R
\name{auto_enter_data}
\alias{auto_enter_data}
\title{A helper function}
\usage{
auto_enter_data(path = path, pattern = ".jpg", start = 1, stop = 3,
  seed = 1, n_frames = 4, n_rows = 5, n_cols = 9,
  lower_hue_threshold = 210, upper_hue_threshold = 230,
  colors = c("empty", "darkred"), img, locs, focal = "NEW",
  case = "N", thresh = 0.25, clean = NA, ordered = NULL,
  lower_saturation_threshold = lower_saturation_threshold,
  lower_luminance_threshold = lower_luminance_threshold,
  upper_luminance_threshold = upper_luminance_threshold,
  border_size = border_size, iso_blur = iso_blur)
}
\arguments{
\item{x}{An object.}
}
\description{
This is the main entry functions
}
