% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_payouts.R
\name{calculate_payouts}
\alias{calculate_payouts}
\title{A function to calculate payouts for RICH economic games using the compiled edgelists}
\usage{
calculate_payouts(path = path, pattern = ".jpg", start = 1,
  stop = 3, game = "GLR", GV = 1, LV = 0.5, KV = 1, RV = -4)
}
\arguments{
\item{path}{Full path to main folder.}

\item{game}{Select the games to merge. "G" for Giving. "L" for Leaving. "R" for Reducing. "GLR" for All. Etc.}

\item{GV}{Value of coins used in giving game.}

\item{LV}{Value of coins used in leaving game.}

\item{KV}{Value of coins kept in reducing game.}

\item{RV}{Value of punishment tokens used in reducing game. Should be negative.}
}
\description{
This function allows you to quickly calculate payouts. See details below.
}
\examples{
\dontrun{
  calculate_payouts(path=path,pattern=".jpg",start=1,stop=3,game="GLR",GV=1,LV=0.5,KV=1,RV=-4)
                   }
}
