% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_classification.R
\name{check_classification}
\alias{check_classification}
\title{A helper function}
\usage{
check_classification(path = "C:/Users/Mind Is Moving/Desktop/RICH/", dat,
  n_frames = 2, n_rows = 4, n_cols = 5, focal = "BS1",
  case = "Friend")
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function to plot infered connections
}
