% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{A helper function}
\usage{
counter(slice, lower_hue_threshold, upper_hue_threshold,
  lower_saturation_threshold = 0.04, lower_luminance_threshold = 0.04,
  upper_luminance_threshold = 0.96, border_size = 8, iso_blur = 0)
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function to check hue
}
