% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{extractor}
\alias{extractor}
\title{A helper function}
\usage{
extractor(image, locations, histogram_balancing = FALSE,
  direction = "backward")
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function
}
