% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\title{A helper function}
\usage{
pre_process(path = path, ID, GID = "Blank", BID = c("A", "B"))
}
\arguments{
\item{x}{An object.}
}
\description{
This is a helper function to pre-process images
}
