% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_color_collisions.R
\name{remove_color_collisions}
\alias{remove_color_collisions}
\title{A function to remove token hues from standardized photos}
\usage{
remove_color_collisions(
  path = path,
  lower_hue_threshold = c(120, 270, 185),
  upper_hue_threshold = c(180, 330, 210),
  rotation_angle = c(-65, 70, 40),
  verbose = FALSE
)
}
\arguments{
\item{path}{Full path to main folder.}

\item{lower_hue_threshold}{A vector of lower hue thresholds for each token color. 
To use three token colors, use: e.g., c(120, 210, 330).}

\item{upper_hue_threshold}{A vector of upper hue thresholds for each token color. 
To use three token colors, use: e.g., c(150, 250, 355).}

\item{rotation_angle}{A vector of angles with which to shift colors inside of the token range. 
Values are matched to the hue thresholds and specified in degree: e.g., c(-65, 70, 40). Calculations are made mod 360.}

\item{verbose}{If TRUE, print which photo is being saved.}
}
\description{
This function allows you to make automatic data entry more accurate by filtering out token color from the photo 
roster. Simply set a path to the main folder. Then supply the token color vectors and shift angles. It is 
recomended to use cool token colors (e.g., green, blue, and purple), and shift all token hues in the standized 
photos twoards warmer yellow, red, and orange hues. All photos in the StandardizedPhotos folder will be updated  
and saved in the PhotosToPrint folder. With token hues removed from the photo-roster, automatic token detection and
classification is much more acurrate, as there a fewer false positives arising from background colors.
}
